+++ src/LoginCallback.tsx
@@ -17,10 +17,10 @@ import OktaError from './OktaError';
 interface LoginCallbackProps {
   errorComponent?: React.ComponentType<{ error: Error }>;
   onAuthResume?: OnAuthResumeFunction;
-  loadingComponent?: React.ReactElement;
+  loadingElement?: React.ReactElement;
 }
 
-const LoginCallback: React.FC<LoginCallbackProps> = ({ errorComponent, loadingComponent = null, onAuthResume }) => { 
+const LoginCallback: React.FC<LoginCallbackProps> = ({ errorComponent, loadingElement = null, onAuthResume }) => { 
   const { oktaAuth, authState } = useOktaAuth();
   const [callbackError, setCallbackError] = React.useState(null);
 
@@ -45,7 +45,7 @@ const LoginCallback: React.FC<LoginCallbackProps> = ({ errorComponent, loadingCo
     return <ErrorReporter error={displayError}/>;
   }
 
-  return loadingComponent;
+  return loadingElement;
 };
 
 export default LoginCallback;